--- a/lib/svg-sprite.js
+++ b/lib/svg-sprite.js
@@ -14,17 +14,19 @@ const path = require('node:path');
 const events = require('node:events');
 const os = require('node:os');
 const { format } = require('node:util');
-const trimStart = require('lodash.trimstart');
-const merge = require('lodash.merge');
-const File = require('vinyl');
-const async = require('async');
-const pretty = require('prettysize');
-const Config = require('./svg-sprite/config.js');
-const Queue = require('./svg-sprite/queue.js');
-const Layouter = require('./svg-sprite/layouter.js');
-const svgo = require('./svg-sprite/transform/svgo.js');
+const importLazy = require('import-lazy');
 const { isFunction, isObject, isPlainObject, zipObject } = require('./svg-sprite/utils/index.js');
-const ArgumentError = require('./svg-sprite/errors/argument-error.js');
+
+const trimStart = importLazy(() => require('lodash.trimstart'))();
+const merge = importLazy(() => require('lodash.merge'))();
+const File = importLazy(() => require('vinyl'))();
+const async = importLazy(() => require('async'))();
+const pretty = importLazy(() => require('prettysize'))();
+const Config = importLazy(() => require('./svg-sprite/config.js'))();
+const Queue = importLazy(() => require('./svg-sprite/queue.js'))();
+const Layouter = importLazy(() => require('./svg-sprite/layouter.js'))();
+const svgo = importLazy(() => require('./svg-sprite/transform/svgo.js'))();
+const ArgumentError = importLazy(() => require('./svg-sprite/errors/argument-error.js'))();
 
 /**
  * SVGSpriter class