```
C:\Users\xmr\Desktop\svg-sprite>git checkout main
Switched to branch 'main'
Your branch is up to date with 'origin/main'.

C:\Users\xmr\Desktop\svg-sprite>hyperfine "node load-time.js"
Benchmark 1: node load-time.js
  Time (mean ± σ):     218.0 ms ±  12.5 ms    [User: 185.0 ms, System: 69.4 ms]
  Range (min … max):   204.4 ms … 243.4 ms    13 runs


C:\Users\xmr\Desktop\svg-sprite>git checkout -
Switched to branch 'xmr/import-lazy'

C:\Users\xmr\Desktop\svg-sprite>hyperfine "node load-time.js"
Benchmark 1: node load-time.js
  Time (mean ± σ):      37.7 ms ±   1.6 ms    [User: 21.6 ms, System: 19.2 ms]
  Range (min … max):    35.6 ms …  44.8 ms    68 runs
```

load-time.js:

```js
'use strict';

console.log('Load time');
console.time('Module');

exports.m = require('.');

console.timeEnd('Module');
```

If/when we switch to ESM, this should be dropped since ESM is inherently faster. But until then, this should improve things.

Unsure if we should use import lazy in the other inner modules too.